refactor rename sort -> sort-menu, sort-menu.js:  extends AbstractComponen
