+++ src/components/sort-menu.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component.js";
 
 const createSortMenuTemplate = () => {
   return (
@@ -10,24 +10,9 @@ const createSortMenuTemplate = () => {
   );
 };
 
-export default class SortMenu {
-  constructor() {
-    this._element = null;
-  }
+export default class SortMenu extends AbstractComponent {
 
   getTemplate() {
     return createSortMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
